week_2_woorkflow_orchestration/docker-requirements.txt
pandas==1.5.2
prefect-sqlalchemy==0.2.4
prefect-gcp[cloud_storage]==0.4.4
protobuf==4.21.11
pyarrow==10.0.1
psycopg2-binary==2.9.5
sqlalchemy==1.4.46